Users/Administrator/Downloads/Programs/Tugas_Individu3_Garis/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.0
google-generativeai==0.3.2